runtime: Fix cast error in print.c on 32-bit systems.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@187889 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libgo/runtime/print.c
@@ -136,7 +136,7 @@ void
 runtime_printpc(void *p __attribute__ ((unused)))
 {
 	runtime_prints("PC=");
-	runtime_printhex((uint64)runtime_getcallerpc(p));
+	runtime_printhex((uint64)(uintptr)runtime_getcallerpc(p));
 }
 
 void
